Fix issue where color prop was being passed to button
Because color is passed to ModeButton it is automatically added to button. To avoid this you can use a wrapper functional component and object destructuring to remove the color prop whilst keeping it accessible to template literal. Unfortunately if this is done directly as a parameter to styled it shows as an unknown component in React inspector. Assigning the wrapper to a constant and passing to styled resolves this.